+++ src/ParserErrorTest.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Bakame\TabularData\HtmlTable;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
+#[CoversClass(ParserError::class)]
 final class ParserErrorTest extends TestCase
 {
     #[Test]